Move integration test
